requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
streamlit==1.31.0
sentence-transformers==2.3.1
transformers==4.37.2
torch==2.2.0
faster-whisper==1.0.3
gtts==2.5.0
faiss-cpu==1.7.4
langdetect==1.0.9
pandas==2.2.0
python-multipart==0.0.9
aiofiles==23.2.1
pydantic==2.6.0
python-dotenv==1.0.1
google-generativeai==0.3.2
numpy==1.26.3
